Add mechanism to create session ids
Create a simple way to create session ids. At the moment, this is using
a pthread spinlock to protect concurrent accesses and it sets a maximum
of 1024 concurrent sessions.

Moreover, add a way to account for errors during initialization. If
initialization of any subsystem has not being succesful we do not fail
at the moment, but we fail in subsequent calls at the subsystem.

Signed-off-by: Babis Chalios <mail@bchalios.io>