Do not use hardcoded networks directory
The imagenet function we were using to load the model was assuming that
the networks were located in the "networks" directory of the CWD of the
application. This is very restrictive and difficult to work with. This
commit changes the functionality to look for networks in (a) a path
defined by an environment variable: `VACCEL_IMAGENET_NETWORKS` or (b) a
default path: `/usr/local/share/imagenet-models/networks`.

Signed-off-by: Babis Chalios <mail@bchalios.io>