Add the generic operation.
VACCEL_GEN_OP is a generic operation where the frontend has a
prototype of the form:

function(output, input, output_len, input_len);

in order to serve as a generic abstraction for any kind of
execution.

The form of input and output parameters is completely agnostic
to the framework. vAccel just forwards a call to a specific
function implemented as a plugin. The actual arguments are unpacked
in the plugin and packed in a small wrapper template in the
frontend (not included in this patchset). For clarity, the wrapper
template should be available as an example to help individuals port
their functions to vAccel.

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>