Add a generic operation dispatcher
In cases where we receive RPC calls (for example from
Firecracker/VirtIO), the arguments arrive in a "packed" format of a
pointer and size tuple. In this case, we need to unpack the arguments
before calling the relevant API function. We add a generic operation
that does exactly that, so that we do not have to do it inside
Firecracker.

Signed-off-by: Babis Chalios <mail@bchalios.io>