Add Rust bindings of vAccel in the repo
Handling the rust bindings of vaccel through a separate repo is very
tedious. In order to fix this issue we add the vaccel-bindings repo (now
called rust-vaccel) as a submodule of this repo.

Adding it as a submodule allows the bindings to be used as a crate from
external Rust projects and at the same time we can test it every time we
build a new version of vAccelRT.

Signed-off-by: Babis Chalios <mail@bchalios.io>