Make building of bindings conditional
Building Rust bindings requires libvaccel to be installed. There is not
easy way around it because rust-vaccel relies on the vaccel.pc package
config file to find the location of libvaccel and this points to the
install prefix.

Make building of bindings conditional to avoid breaking the build of
the core library.

Signed-off-by: Babis Chalios <mail@bchalios.io>