vAccel-exec: refactor genop to vaccel-exec
is a generic operation where the frontend has a
prototype of the form:

`function(lib, symbol, input, input_len, output, output_len);`

in order to serve as a generic abstraction for any kind of
execution.

The form of input and output parameters is of type

`struct vaccel_arg {
	u32 len;
	u8 *buf;
}`

vAccel forwards a call to the specified `symbol`, found in the
specified `lib`, with all the arguments serialized in arrays of
the above structure (read & write).

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>