Introduce vAccel resources
A vAccel resource represents conceptually objects that can be consumed
by vAccel operations. A resource can be created/destroyed and
(un)registered with a sesssion. vAccel operations can use resources
associated (registered) with the session within which they are
operating.

As part of this commit we add a generic identifier pool for generating
ids for resources so that we do not duplicate the code for handling
object identifiers. We use such pools for both vAccel resources and
sessions.

Signed-off-by: Babis Chalios <mail@bchalios.io>