Create `vaccel_file` data structure to assist with resources
Many of the resources we will be handling need to handle an underlying
file on persistent storage. We create `struct vaccel_file` as a helper
data structure to handle reading such a resource from the disk or
creating it from raw data and persisting it as a resource when needed.

Signed-off-by: Babis Chalios <mail@bchalios.io>