Do not use a resource for persisting a file
The API for persisting a vaccel file was receiving a pointer to a
vAccel resource with the intention to use its rundir for persisting the
file. This is very restricting, since it does not allow us to save the
file in subdirectories of the rundir.

Change the API so that it receives a path where to persist the file.

Signed-off-by: Babis Chalios <mail@bchalios.io>