Correct handling of file persisting and removing files
When we persist a file in the file-system we mmap it to get a pointer to
its data. When destroying such a file we were using `free` to deallocate
the memory instead of `munmap` leading to crashes.

Moreover, when we had a vAccel file created from a buffer in memory,
e.g. that is the case when we receive a file over gRPC, and we persist
that file, we are writing the memory in the file and then mmaping it,
without though deallocating the initial memory pointer, leading to a
memory leak.

Finally, when cleaning up a TensorFlow SavedModel resource, we were not
removing the `variables` directory from the resource rundir, which was
obstructing cleaning up the rundir later on.

Signed-off-by: Babis Chalios <mail@bchalios.io>