Do not own memory for SavedModel files
When creating a SavedModel from in-memory data we were taking ownership
of said data. This is problematic, because we do not know how to
deallocate the memory.

Instead of requiring the user to pass us a callback to use for
deallocating, we just do not take ownership of the pointer. This is
fine. We only own the data when we persist such a file, really, but then
we substitute the pointer with the mmaped file, which we do know how to
handle.

Also, update the example of handling SavedModel to cover all of the
relevant API.

Signed-off-by: Babis Chalios <mail@bchalios.io>