Fix noop example when called from virtio/vsock plugin
When we call the noop example from a virtio/vsock plugin,
the argument check enters the error-if clause as one of
the intermediate buffers is not null-ified.

We still have to get the argument list, to keep a consistent
function signature for genop. This causes the compiler to issue
a warning about the unused vars. Add a simple print to avoid that.

Signed-off-by: Anastassios nanos <ananos@nubificus.co.uk>