resources: randomize object filename
We hit a really weird issue when running exec with resource over vsock.
Essentially, any consecutive calls to the same resource file (even if
the session is closed and the resource is freed) keep using the data of
the first resource passed.

For now, we can just randomize the name of the file, but this is not a
permanent solution.

Wild guess: this points to some weird cache issue either with mmap() of
with the rust<->C mmap() call. Major FIXME!!

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>