API: add a wrapper API operation for OpenCV functions
Essentially, provide an API call to facilitate OpenCV integration.
`vaccel_opencv()` is a `vaccel_genop()`. We only use a different
name for clarity (and as a placeholder, in case we would like to
change the API).

Along with the rest of the supported OpenCV operations
(currently WiP), we will be able to provide end-to-end
OpenCV use-cases that off-load cpu-intensive operations
to vAccel plugins.

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>