catch2 fix
for some reason - I didn't catch this one but I was using catch3 the entire time - where the philosophy of the drag and drop header has changed. This commit should address that and revert back to an older version of catch which only uses a header file and aims to speed up the build process since it is currently taking 6 minutes + on some of the runners.