feat: Introduce `vaccel_session_resource_by_type()` function
The new function lets the user get access to a vaccel resource
that matches with the input `type` and is registered to the input
vaccel session `sess`. Unless a matching resource is registered to
the calling session, the function will return `VACCEL_EINVAL`.

There's also a new testing session to test all scenarios of the
new function's usage.

PR: https://github.com/nubificus/vaccel/pull/151
Signed-off-by: Ilias Lagomatis <ilago@nubificus.co.uk>
Reviewed-by: Kostis Papazafeiropoulos <papazof@gmail.com>
Approved-by: Kostis Papazafeiropoulos <papazof@gmail.com>