feat(examples): Introduce additional options for examples
Allow examples (for now `classify.c`) to specify the model
against which inference is going to be done in the plugin.
For now, we assume that the model is loaded as a resource,
with type `VACCEL_RESOURCE_MODEL` and the plugin pops the first
available resource with this type to use it.

This is useful in our TVM example.

Add this as an optional argument in meson:
`meson setup -Dexamples-model=enabled --reconfigure build`

Additionally, add an extra parameter for profiling:

`meson setup -Dexamples-profile=enabled --reconfigure build`

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>