feat(plugins): Resolve symbols early
Until now, we rely on RTLD_LAZY to resolve plugin-specific symbols.
Using TVM, there were symbols that could not be resolved at runtime,
and need to be resolved early on. Apparently, we should have defined
the flag regardless:

`If neither RTLD_GLOBAL nor RTLD_LOCAL are specified, then the
default behavior is unspecified.`
https://pubs.opengroup.org/onlinepubs/009696799/functions/dlopen.html

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>