feat(examples): Add optional model arg in image ops' examples
Enable selection of an inference model at runtime via an optional
argument for image ops' examples. For now, we assume that the model is
loaded as a resource with type `VACCEL_RESOURCE_MODEL` and the
respective plugin must pop the first available resource with this type
to use it.

PR: https://github.com/nubificus/vaccel/pull/164
Signed-off-by: Kostis Papazafeiropoulos <papazof@gmail.com>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>