fix(plugin): Introduce proper check for extra fields in version
When we check the version compatibility between the core library and
plugins, we compare the version string against the plugin version
string. We use `strcmp` to do that, and its behavior is undefined
when we pass a `NULL` ptr. This check verifies that the strings passed
to `strcmp` are not `NULL`.

PR: https://github.com/nubificus/vaccel/pull/174
Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Reviewed-by: Kostis Papazafeiropoulos <papazof@gmail.com>
Approved-by: Kostis Papazafeiropoulos <papazof@gmail.com>