feat(resource): Add plugin private data attribute
Also, the attribute is coming along with a state field, which
can be a value between empty, init, need_free. This way, the
release function will know whether it must free the memory space
pointed by `plugin_priv` or not. Furthermore, the plugin will
know if its resource private data should be initialized or not.

Signed-off-by: Ilias Lagomatis <ilago@nubificus.co.uk>