refactor: Associate plugin discovery with session
Find and set a plugin per session instead of discovering a plugin upon
operation execution.

This changes the previous behavior, where all loaded plugins would be
searched for operation support. The plugin is selected upon session
creation/update, based on the provided hint, regardless of the
operations that might be executed using the session.

Signed-off-by: Kostis Papazafeiropoulos <papazof@gmail.com>