refactor: Introduce new struct to handle resource registration
Introduce `resource_registration` struct to handle resource
registration. The new struct replaces `struct registered_resource` and
stores pointers for both the registered resource and the related
session. The API for `struct resource_registration` is thread-safe and
the registration is added to lists in both session and resource structs.

This enables resource unregistration upon release and removes the
requirement for ordered cleanup, ie. resources or sessions can be
cleaned up in any order, even if there are registered resources (not
unregistered by the user).

PR: https://github.com/nubificus/vaccel/pull/193
Signed-off-by: Kostis Papazafeiropoulos <papazof@gmail.com>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>